Refactored code into specific .ts files for better structure
